backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
cryptography==43.0.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
requests==2.32.3
python-multipart==0.0.12
beautifulsoup4==4.12.3
pytrends==4.9.2
google-api-python-client==2.149.0
